refactor: toEntity 형식으로 객체 생성
